replace module_path with current_list_dir
module_path contains a semicolon seperated list of paths to search in.
This is why cmake isn't able to find the `safeConfig.cmake.in` in my
case. The `CMAKE_CURRENT_LIST_DIR` holds the value of the current
directory the cmake file is executed in. This means, that the
`safeConfig.cmake.in` needs to reside in the same directory as the
InstallTarget file, but this is the case and therefore shouldn't be a
problem.